+++ CHANGES.md
@@ -1,5 +1,13 @@
 [unreleased]
 
+#### 0.6.0 / 2021-03-13
+* remove constructor
+* add some tests
+* add filter `gu_get_repo_api`
+* add filter `gu_get_auth_header`
+* add filter `gu_post_get_credentials`
+* add filter `gu_parse_release_asset`
+
 #### 0.5.0 / 2021-03-12
 * de-anonymize hooks
 * add filters for language pack processing